+++ packages/terra-form-select/src/shared/_Tag.jsx
@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { useRef } from 'react';
 import PropTypes from 'prop-types';
 import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
+import { injectIntl } from 'react-intl';
 import styles from './_Tag.module.scss';
 
 const cx = classNamesBind.bind(styles);
@@ -19,17 +20,75 @@ const propTypes = {
    * The value of the tag.
    */
   value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+  /**
+   * Specifies whether the tag is disabled.
+   */
+  disabled: PropTypes.bool,
+  /**
+   * @private
+   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
+   */
+  intl: PropTypes.shape({ formatMessage: PropTypes.func }).isRequired,
+  /**
+   * Ref object for accessing the underlying input element of the tag component.
+   */
+  inputRef: PropTypes.shape({
+    focus: PropTypes.instanceOf(Element),
+  }),
+  /**
+   * Specifies whether the input focus is set to true or false.
+   * Default is false.
+   */
+  isInputFocused: PropTypes.bool,
 };
 
 /* eslint-disable jsx-a11y/no-static-element-interactions */
-const Tag = ({ children, onDeselect, value }) => {
+const Tag = ({
+  children, onDeselect, value, disabled, intl, inputRef, isInputFocused,
+}) => {
   const theme = React.useContext(ThemeContext);
+  const tagRef = useRef(null);
+
+  const handleKeyPress = (event) => {
+    if ((event.key === 'Enter' || event.key === 'Backspace') && !disabled) {
+      event.stopPropagation();
+      onDeselect(value);
+      const previousLi = tagRef.current.previousElementSibling;
+      if (previousLi) {
+        const deselectElement = previousLi.querySelector(':scope > :nth-child(2)');
+        if (deselectElement) {
+          deselectElement.focus();
+        }
+      } else {
+        const nextLi = tagRef.current.nextElementSibling;
+        if (nextLi) {
+          const nextFocusableElement = nextLi.querySelector(':scope > :nth-child(2)');
+          if (nextFocusableElement) {
+            nextFocusableElement.focus();
+            return;
+          }
+        }
+        inputRef.focus();
+      }
+    }
+  };
+
+  const attributes = isInputFocused ? { role: 'presentation' }
+    : { role: 'button', 'aria-label': intl.formatMessage({ id: 'Terra.form.select.deselect' }, { text: children }) };
   return (
-    <li className={cx('tag', theme.className)}>
+    <li className={cx('tag', theme.className)} ref={tagRef}>
       <span className={cx('display')}>
         {children}
       </span>
-      <span className={cx('deselect')} onClick={() => { onDeselect(value); }} role="presentation">
+      <span
+        id={`terra-tag-deselect-${value}`}
+        onKeyDown={handleKeyPress}
+        className={cx('deselect')}
+        onClick={() => { if (!disabled) onDeselect(value); }}
+        tabIndex={!disabled ? 0 : -1}
+        role="button"
+        {...attributes}
+      >
         <span className={cx('icon')} />
       </span>
     </li>
@@ -38,4 +97,4 @@ const Tag = ({ children, onDeselect, value }) => {
 
 Tag.propTypes = propTypes;
 
-export default Tag;
+export default injectIntl(Tag);